+++ src/App.js
@@ -22,7 +22,7 @@ import Error503 from './errorPages/Error503';
 import Step1Page from './pages/Step1Page';
 import Step2Page from './pages/Step2Page';
 import Step3Page from './pages/Step3Page';
-import Step4 from './pages/Step4';
+import Step4Page from './pages/Step4Page';
 import Step5 from './pages/Step5';
 
 
@@ -46,7 +46,7 @@ function App() {
       <Route path="/step1page" element={<Step1Page language={language} onChangeLanguage={setLanguage} />} />
       <Route path="/step2page" element={<Step2Page language={language} onChangeLanguage={setLanguage} />} />
       <Route path="/step3page" element={<Step3Page language={language} onChangeLanguage={setLanguage} />} />
-      <Route path="/step4" element={<Step4 language={language} onChangeLanguage={setLanguage} />} />
+      <Route path="/step4Page" element={<Step4Page language={language} onChangeLanguage={setLanguage} />} />
       <Route path="/step5" element={<Step5 language={language} onChangeLanguage={setLanguage} />} />
 
       {/* 어드민 라우트 */}
